--- a/core/amber/build.sbt
+++ b/core/amber/build.sbt
@@ -189,8 +189,8 @@ libraryDependencies += "com.typesafe.scala-logging" %% "scala-logging" % "3.9.2"
 // https://mvnrepository.com/artifact/org.scalactic/scalactic
 libraryDependencies += "org.scalactic" %% "scalactic" % "3.2.2"
 
-// https://mvnrepository.com/artifact/com.github.tototoshi/scala-csv
-libraryDependencies += "com.github.tototoshi" %% "scala-csv" % "1.3.6"
+// https://mvnrepository.com/artifact/com.univocity/univocity-parsers
+libraryDependencies += "com.univocity" % "univocity-parsers" % "2.9.1"
 
 // https://mvnrepository.com/artifact/com.konghq/unirest-java
 libraryDependencies += "com.konghq" % "unirest-java" % "3.11.11"