--- a/core/amber/src/main/scala/edu/uci/ics/texera/workflow/operators/source/scan/csv/CSVScanSourceOpDesc.scala
+++ b/core/amber/src/main/scala/edu/uci/ics/texera/workflow/operators/source/scan/csv/CSVScanSourceOpDesc.scala
@@ -6,16 +6,12 @@ import com.kjetland.jackson.jsonSchema.annotations.JsonSchemaTitle
 import edu.uci.ics.amber.engine.operators.OpExecConfig
 import edu.uci.ics.texera.workflow.common.operators.ManyToOneOpExecConfig
 import edu.uci.ics.texera.workflow.common.tuple.schema.AttributeTypeUtils.inferSchemaFromRows
-import edu.uci.ics.texera.workflow.common.tuple.schema.{
-  Attribute,
-  AttributeType,
-  OperatorSchemaInfo,
-  Schema
-}
+import edu.uci.ics.texera.workflow.common.tuple.schema.{Attribute, AttributeType, OperatorSchemaInfo, Schema}
 import edu.uci.ics.texera.workflow.operators.source.scan.ScanSourceOpDesc
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize
+import com.univocity.parsers.csv.{CsvFormat, CsvParser, CsvParserSettings}
 
-import java.io.IOException
+import java.io.{File, FileInputStream, IOException, InputStreamReader}
 import scala.jdk.CollectionConverters.asJavaIterableConverter
 
 class CSVScanSourceOpDesc extends ScanSourceOpDesc {
@@ -61,40 +57,50 @@ class CSVScanSourceOpDesc extends ScanSourceOpDesc {
     if (filePath.isEmpty) {
       return null
     }
-    implicit object CustomFormat extends DefaultCSVFormat {
-      override val delimiter: Char = customDelimiter.get.charAt(0)
-    }
-    var reader: CSVReader = CSVReader.open(filePath.get)(CustomFormat)
-    val firstRow: Array[String] = reader.iterator.next().toArray
-    reader.close()
+    val inputReader = new InputStreamReader(new FileInputStream(
+      new File(filePath.get)))
 
-    // reopen the file to read from the beginning
-    reader = CSVReader.open(filePath.get)(CustomFormat)
+    val csvFormat = new CsvFormat()
+    csvFormat.setDelimiter(customDelimiter.get.charAt(0))
+    val csvSetting = new CsvParserSettings()
+    csvSetting.setMaxCharsPerColumn(-1)
+    csvSetting.setFormat(csvFormat)
+    val csvParser = new CsvParser(csvSetting)
+    csvParser.beginParsing(inputReader)
 
-    val startOffset = offset.getOrElse(0) + (if (hasHeader) 1 else 0)
-    val endOffset =
-      startOffset + limit.getOrElse(INFER_READ_LIMIT).min(INFER_READ_LIMIT)
-    val attributeTypeList: Array[AttributeType] = inferSchemaFromRows(
-      reader.iterator
-        .slice(startOffset, endOffset)
-        .map(seq => seq.toArray)
-    )
+    csvSetting.setNumberOfRowsToSkip()
 
-    reader.close()
+    val header =
+      if (hasHeader) Some(csvParser.parseNext()) else None
 
-    // build schema based on inferred AttributeTypes
-    Schema.newBuilder
-      .add(
-        firstRow.indices
-          .map((i: Int) =>
-            new Attribute(
-              if (hasHeader) firstRow.apply(i) else "column-" + (i + 1),
-              attributeTypeList.apply(i)
-            )
-          )
-          .asJava
-      )
-      .build
+    // skip offset
+//    csvParser.
+
+
+//    val startOffset = offset.getOrElse(0) + (if (hasHeader) 1 else 0)
+//    val endOffset =
+//      startOffset + limit.getOrElse(INFER_READ_LIMIT).min(INFER_READ_LIMIT)
+//    val attributeTypeList: Array[AttributeType] = inferSchemaFromRows(
+//      reader.iterator
+//        .slice(startOffset, endOffset)
+//        .map(seq => seq.toArray)
+//    )
+//
+//    reader.close()
+//
+//    // build schema based on inferred AttributeTypes
+//    Schema.newBuilder
+//      .add(
+//        firstRow.indices
+//          .map((i: Int) =>
+//            new Attribute(
+//              if (hasHeader) firstRow.apply(i) else "column-" + (i + 1),
+//              attributeTypeList.apply(i)
+//            )
+//          )
+//          .asJava
+//      )
+//      .build
   }
 
 }