This PR switchs the CSV parser library from [scala-csv](https://github.com/tototoshi/scala-csv) to [univocity-parser](https://github.com/uniVocity/univocity-parsers). 
This is because `scala-csv` is very very slow for csv data with large rows. This makes the system completely unresponsive because of doing schema inference for CSV files.

Here's some simple comparison statistics of CSV parser libraries. The CSV file contains 2 HTML pages in each record.
|                      | 100 records                    | 1000 records                    |
|----------------------|-----------------------------|------------------------------|
| scala-csv (current)            | 250 **seconds** (estimated) | 2500 **seconds** (estimated) |
| apache-commons-csv   | 294 ms                      | 1865 ms                      |
| **univocity-parser-csv** | 153 ms                      | 894ms                        |

We choose univocity-parser-csv because it's fastest. Here's another online comparsion as a reference: https://github.com/uniVocity/csv-parsers-comparison


Moreover, the speed of schema inference is also tested. We compare our own implementation vs the schema inference implementation in [spark-csv](https://github.com/databricks/spark-csv/blob/master/src/main/scala/com/databricks/spark/csv/util/InferSchema.scala). 

|                      | 1000 records |
|----------------------|--------------|
| our schema inference | 31 ms        |
| Spark-CSV inference  | 17 ms        |

Result shows spark-csv's implementation is faster than our own implementation, however, since this is not a bottleneck, we stick with our current implementation.