@zuozhiw 2022-05-26
> LGTM.
> 
> 1. We might also deprecate other places where uses this library.
> 2. Please test extensively with our tweet datasets and broken datasets (for errors).
> 3. Please consider increase infer limit as we have a faster parser now. or use dynamic method as you described.

1. I'll do this in future PRs 
2. Yes I have tested many CSV files in the production envrionment using the new code
3. I'll implement the dynamic method in future PRs